Cleanup instrumentation tests
 * Adds a similar base class that contains boilerplate code for writing instrumentation tests
